requirements.txt
django
djangorestframework
Pillow